+++ demo/demo.jsx
@@ -1,17 +1,18 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import range from 'lodash.range';
 import CalendarHeatmap from '../src';
-import { shiftDate } from '../src/dateHelpers';
+import { dateNDaysAgo, shiftDate, getRange } from '../src/helpers';
 
 const today = new Date();
+const date100DaysBefore = dateNDaysAgo(100);
+const date31DaysBefore = dateNDaysAgo(31);
 
 function getRandomInt(min, max) {
   return Math.floor(Math.random() * (max - min + 1)) + min;
 }
 
 function generateRandomValues(count, date = today) {
-  return range(count).map((index) => {
+  return getRange(count).map((index) => {
     return {
       date: shiftDate(date, -index),
       count: getRandomInt(1, 3),
@@ -92,6 +93,7 @@ class Demo extends React.Component {
               titleForValue={customTitleForValue}
               tooltipDataAttrs={customTooltipDataAttrs}
               onClick={customOnClick}
+              monthLabels={['01', '02', '03', '04', '05', '06', '07', '08', '09', '10', '11', '12']}
             />
           </div>
         </div>
@@ -109,120 +111,103 @@ class Demo extends React.Component {
           name="values"
           example="[{ date: '2016-01-01', count: 6 }]"
           description="Required array of objects which each have a date property, which can be a Date object, parseable string, or millisecond timestamp."
-        >
-        </DemoItem>
+        />
 
         <DemoItem
-          name="numDays"
-          example="200"
-          description="Time span in days."
-        >
-        </DemoItem>
+          name="startDate"
+          example="new Date()"
+          description="Start of date range - a Date object, parseable string, or millisecond timestamp."
+        />
 
         <DemoItem
           name="endDate"
           example="new Date()"
           description="End of date range - a Date object, parseable string, or millisecond timestamp."
-        >
-        </DemoItem>
+        />
 
         <DemoItem
           name="showMonthLabels"
           example="true"
           description="Toggle for removing month labels."
-        >
-          <div className="row">
-            <CalendarHeatmap
-              values={randomValues}
-              showMonthLabels={false}
-            />
-          </div>
-        </DemoItem>
+        />
+
+        <DemoItem
+          name="showWeekdayLabels"
+          example="true"
+          description="Toggle for removing weekday labels."
+        />
 
         <DemoItem
           name="showOutOfRangeDays"
           example="false"
           description="Toggle display of extra days in week that are past endDate and before beginning of range."
-        >
-          <div className="row">
-            <CalendarHeatmap
-              values={randomValues}
-              showOutOfRangeDays={true}
-            />
-          </div>
-        </DemoItem>
+        />
 
         <DemoItem
           name="horizontal"
           example="true"
-          description="Whether to orient horizontally or vertically. Can be used in combination with numDays/endDate to show just the current month."
-        >
-          <div className="row">
-            <div className="col-xs-6">
-              <CalendarHeatmap
-                values={randomValues}
-                numDays={100}
-                horizontal={false}
-              />
-            </div>
-            <div className="col-xs-6">
-              <CalendarHeatmap
-                values={randomValues}
-                numDays={31}
-                horizontal={false}
-                showMonthLabels={false}
-              />
-            </div>
-          </div>
-        </DemoItem>
+          description="Whether to orient horizontally or vertically. Can be used in combination with startDate/endDate to show just the current month."
+        />
 
         <DemoItem
           name="gutterSize"
           example="1"
           description="Size of gutters relative to squares."
-        >
-          <div className="row">
-            <CalendarHeatmap
-              values={randomValues}
-              gutterSize={2}
-            />
-          </div>
-        </DemoItem>
+        />
 
         <DemoItem
           name="onClick"
           example="(value) => { alert(value) }"
           description="Callback to invoke when a square is clicked."
-        >
-        </DemoItem>
+        />
+
+        <DemoItem
+          name="onMouseOver"
+          example="(event, value) => { console.log(event, value) }"
+          description="Callback to invoke when mouse pointer is over a square."
+        />
+
+        <DemoItem
+          name="onMouseLeave"
+          example="(event, value) => { console.log(event, value) }"
+          description="Callback to invoke when mouse pointer is left a square."
+        />
 
         <DemoItem
           name="titleForValue"
           example="(value) => `Date is ${value.date}`"
           description="Function to determine each square's title attribute, for generating 3rd party hover tooltips (may also need to configure tooltipDataAttrs)."
-        >
-        </DemoItem>
+        />
 
         <DemoItem
           name="tooltipDataAttrs"
           example="{ 'data-toggle': 'tooltip' } or (value) => ({ 'data-tooltip': `Tooltip: ${value}` })"
           description="Sets data attributes for all squares, for generating 3rd party hover tooltips (this demo uses bootstrap tooltips)."
-        >
-        </DemoItem>
+        />
 
         <DemoItem
           name="classForValue"
           example="(value) => (value.count > 0 ? 'blue' : 'white')"
           description="Callback for determining CSS class to apply to each value."
-        >
-        </DemoItem>
+        />
+
+        <DemoItem
+          name="monthLabels"
+          example="['01', '02', ..., '12']"
+          description="An array with 12 strings representing the text from janurary to december"
+        />
+
+        <DemoItem
+          name="weekdayLabels"
+          example="['Sun', 'Mon', ..., 'Sat']"
+          description="An array with 7 strings representing the text from Sunday to Saturday"
+        />
 
         <DemoItem
           name="transformDayElement"
           example="(element, value, index) => React.cloneElement(element, {title: `${value.date}, ${index}`)"
           description="A function to further transform generated svg element for a single day, can be used to attach event handlers, add tooltips and more"
-        >
-        </DemoItem>
+        />
 
         <hr />
         <div className="text-xs-center my-3">