+++ src/index.jsx
@@ -3,46 +3,71 @@ import PropTypes from 'prop-types';
 import range from 'lodash.range';
 import reduce from 'lodash.reduce';
 import {Tooltip,OverlayTrigger} from 'react-bootstrap';
-import { DAYS_IN_WEEK, MILLISECONDS_IN_ONE_DAY, MONTH_LABELS } from './constants';
-import { shiftDate, getBeginningTimeForDate, convertToDate } from './dateHelpers';
+import { DAYS_IN_WEEK, MILLISECONDS_IN_ONE_DAY, DAY_LABELS, MONTH_LABELS } from './constants';
+import { dateNDaysAgo, shiftDate, getBeginningTimeForDate, convertToDate, getRange } from './helpers';
 
 const SQUARE_SIZE = 10;
 const MONTH_LABEL_GUTTER_SIZE = 4;
+const CSS_PSEDUO_NAMESPACE = 'react-calendar-heatmap-';
 
 class CalendarHeatmap extends React.Component {
   constructor(props) {
     super(props);
 
+    this.latestProps = props;
+
     this.state = {
-      valueCache: this.getValueCache(props.values),
+      valueCache: this.getValueCache(this.latestProps.values),
     };
   }
 
   componentWillReceiveProps(nextProps) {
+    this.latestProps = nextProps;
+
     this.setState({
-      valueCache: this.getValueCache(nextProps.values),
+      valueCache: this.getValueCache(this.latestProps.values),
     });
   }
 
+  getDateDifferenceInDays() {
+    const { startDate, numDays } = this.latestProps;
+    if (numDays) {
+      // eslint-disable-next-line no-console
+      console.warn('numDays is a deprecated prop. It will be removed in the next release. Consider using the startDate prop instead.');
+      return numDays;
+    }
+    const timeDiff = this.getEndDate() - convertToDate(startDate);
+    return Math.ceil(timeDiff / MILLISECONDS_IN_ONE_DAY);
+  }
+
   getSquareSizeWithGutter() {
-    return SQUARE_SIZE + this.props.gutterSize;
+    return SQUARE_SIZE + this.latestProps.gutterSize;
   }
 
   getMonthLabelSize() {
-    if (!this.props.showMonthLabels) {
+    if (!this.latestProps.showMonthLabels) {
       return 0;
-    } else if (this.props.horizontal) {
+    } else if (this.latestProps.horizontal) {
       return SQUARE_SIZE + MONTH_LABEL_GUTTER_SIZE;
     }
     return 2 * (SQUARE_SIZE + MONTH_LABEL_GUTTER_SIZE);
   }
 
+  getWeekdayLabelSize() {
+    if (!this.latestProps.showWeekdayLabels) {
+      return 0;
+    } else if (this.latestProps.horizontal) {
+      return 30;
+    }
+    return SQUARE_SIZE * 1.5;
+  }
+
   getStartDate() {
-    return shiftDate(this.getEndDate(), -this.props.numDays + 1); // +1 because endDate is inclusive
+    return shiftDate(this.getEndDate(), -this.getDateDifferenceInDays() + 1); // +1 because endDate is inclusive
   }
 
   getEndDate() {
-    return getBeginningTimeForDate(convertToDate(this.props.endDate));
+    return getBeginningTimeForDate(convertToDate(this.latestProps.endDate));
   }
 
   getStartDateWithEmptyDays() {
@@ -58,7 +83,7 @@ class CalendarHeatmap extends React.Component {
   }
 
   getWeekCount() {
-    const numDaysRoundedToWeek = this.props.numDays + this.getNumEmptyDaysAtStart() + this.getNumEmptyDaysAtEnd();
+    const numDaysRoundedToWeek = this.getDateDifferenceInDays() + this.getNumEmptyDaysAtStart() + this.getNumEmptyDaysAtEnd();
     return Math.ceil(numDaysRoundedToWeek / DAYS_IN_WEEK);
   }
 
@@ -67,21 +92,21 @@ class CalendarHeatmap extends React.Component {
   }
 
   getWidth() {
-    return (this.getWeekCount() * this.getSquareSizeWithGutter()) - this.props.gutterSize;
+    return (this.getWeekCount() * this.getSquareSizeWithGutter()) - (this.latestProps.gutterSize - this.getWeekdayLabelSize());
   }
 
   getHeight() {
-    return this.getWeekWidth() + (this.getMonthLabelSize() - this.props.gutterSize);
+    return this.getWeekWidth() + (this.getMonthLabelSize() - this.latestProps.gutterSize) + this.getWeekdayLabelSize();
   }
 
   getValueCache(values) {
-    return reduce(values, (memo, value) => {
+    return values.reduce((memo, value) => {
       const date = convertToDate(value.date);
       const index = Math.floor((date - this.getStartDateWithEmptyDays()) / MILLISECONDS_IN_ONE_DAY);
       memo[index] = {
         value,
-        className: this.props.classForValue(value),
-        title: this.props.titleForValue ? this.props.titleForValue(value) : null,
+        className: this.latestProps.classForValue(value),
+        title: this.latestProps.titleForValue ? this.latestProps.titleForValue(value) : null,
         tooltipDataAttrs: this.getTooltipDataAttrsForValue(value),
       };
       return memo;
@@ -99,14 +124,14 @@ class CalendarHeatmap extends React.Component {
     if (this.state.valueCache[index]) {
       return this.state.valueCache[index].className;
     }
-    return this.props.classForValue(null);
+    return this.latestProps.classForValue(null);
   }
 
   getTitleForIndex(index) {
     if (this.state.valueCache[index]) {
       return this.state.valueCache[index].title;
     }
-    return this.props.titleForValue ? this.props.titleForValue(null) : null;
+    return this.latestProps.titleForValue ? this.latestProps.titleForValue(null) : null;
   }
 
   getTooltipDataAttrsForIndex(index) {
@@ -117,7 +142,7 @@ class CalendarHeatmap extends React.Component {
   }
 
   getTooltipDataAttrsForValue(value) {
-    const { tooltipDataAttrs } = this.props;
+    const { tooltipDataAttrs } = this.latestProps;
 
     if (typeof tooltipDataAttrs === 'function') {
       return tooltipDataAttrs(value);
@@ -126,42 +151,62 @@ class CalendarHeatmap extends React.Component {
   }
 
   getTransformForWeek(weekIndex) {
-    if (this.props.horizontal) {
+    if (this.latestProps.horizontal) {
       return `translate(${weekIndex * this.getSquareSizeWithGutter()}, 0)`;
     }
     return `translate(0, ${weekIndex * this.getSquareSizeWithGutter()})`;
   }
 
-  getTransformForMonthLabels() {
-    if (this.props.horizontal) {
-      return null;
-    }
-    return `translate(${this.getWeekWidth() + MONTH_LABEL_GUTTER_SIZE}, 0)`;
-  }
-
-  getTransformForAllWeeks() {
-    if (this.props.horizontal) {
-      return `translate(0, ${this.getMonthLabelSize()})`;
+  getTransformForWeekdayLabels() {
+    if (this.latestProps.horizontal) {
+      return `translate(${SQUARE_SIZE}, ${this.getMonthLabelSize()})`;
     }
     return null;
   }
 
+  getTransformForMonthLabels() {
+    if (this.latestProps.horizontal) {
+      return `translate(${this.getWeekdayLabelSize()}, 0)`;
+    }
+    return `translate(${this.getWeekWidth() + MONTH_LABEL_GUTTER_SIZE}, ${this.getWeekdayLabelSize()})`;
+  }
+
+  getTransformForAllWeeks() {
+    if (this.latestProps.horizontal) {
+      return `translate(${this.getWeekdayLabelSize()}, ${this.getMonthLabelSize()})`;
+    }
+    return `translate(0, ${this.getWeekdayLabelSize()})`;
+  }
+
   getViewBox() {
-    if (this.props.horizontal) {
+    if (this.latestProps.horizontal) {
       return `0 0 ${this.getWidth()} ${this.getHeight()}`;
     }
     return `0 0 ${this.getHeight()} ${this.getWidth()}`;
   }
 
   getSquareCoordinates(dayIndex) {
-    if (this.props.horizontal) {
+    if (this.latestProps.horizontal) {
       return [0, dayIndex * this.getSquareSizeWithGutter()];
     }
     return [dayIndex * this.getSquareSizeWithGutter(), 0];
   }
 
+  getWeekdayLabelCoordinates(dayIndex) {
+    if (this.latestProps.horizontal) {
+      return [
+        0,
+        ((dayIndex + 1) * SQUARE_SIZE) + (dayIndex * this.latestProps.gutterSize),
+      ];
+    }
+    return [
+      (dayIndex * SQUARE_SIZE) + (dayIndex * this.latestProps.gutterSize),
+      SQUARE_SIZE,
+    ];
+  }
+
   getMonthLabelCoordinates(weekIndex) {
-    if (this.props.horizontal) {
+    if (this.latestProps.horizontal) {
       return [
         weekIndex * this.getSquareSizeWithGutter(),
         this.getMonthLabelSize() - MONTH_LABEL_GUTTER_SIZE,
@@ -175,14 +220,26 @@ class CalendarHeatmap extends React.Component {
   }
 
   handleClick(value) {
-    if (this.props.onClick) {
-      this.props.onClick(value);
+    if (this.latestProps.onClick) {
+      this.latestProps.onClick(value);
+    }
+  }
+
+  handleMouseOver(e, value) {
+    if (this.latestProps.onMouseOver) {
+      this.latestProps.onMouseOver(e, value);
+    }
+  }
+
+  handleMouseLeave(e, value) {
+    if (this.latestProps.onMouseLeave) {
+      this.latestProps.onMouseLeave(e, value);
     }
   }
 
   renderSquare(dayIndex, index) {
-    const indexOutOfRange = index < this.getNumEmptyDaysAtStart() || index >= this.getNumEmptyDaysAtStart() + this.props.numDays;
-    if (indexOutOfRange && !this.props.showOutOfRangeDays) {
+    const indexOutOfRange = index < this.getNumEmptyDaysAtStart() || index >= this.getNumEmptyDaysAtStart() + this.getDateDifferenceInDays();
+    if (indexOutOfRange && !this.latestProps.showOutOfRangeDays) {
       return null;
     }
     const [x, y] = this.getSquareCoordinates(dayIndex);
@@ -198,40 +255,55 @@ class CalendarHeatmap extends React.Component {
           title={this.getTitleForIndex(index)}
           className={this.getClassNameForIndex(index)}
           onClick={this.handleClick.bind(this, value)}
+          onMouseOver={e => this.handleMouseOver(e, value)}
+          onMouseLeave={e => this.handleMouseLeave(e, value)}
+          {...this.getTooltipDataAttrsForIndex(index)}
         />
       </OverlayTrigger>
     );
-    const transformDayElement = this.props.transformDayElement;
+    const { transformDayElement } = this.latestProps;
     return transformDayElement ? transformDayElement(rect, value, index) : rect;
   }
 
   renderWeek(weekIndex) {
     return (
-      <g key={weekIndex} transform={this.getTransformForWeek(weekIndex)}>
-        {range(DAYS_IN_WEEK).map(dayIndex => this.renderSquare(dayIndex, (weekIndex * DAYS_IN_WEEK) + dayIndex))}
+      <g key={weekIndex} transform={this.getTransformForWeek(weekIndex)} className={`${CSS_PSEDUO_NAMESPACE}week`}>
+        {getRange(DAYS_IN_WEEK).map(dayIndex => this.renderSquare(dayIndex, (weekIndex * DAYS_IN_WEEK) + dayIndex))}
       </g>
     );
   }
 
   renderAllWeeks() {
-    return range(this.getWeekCount()).map(weekIndex => this.renderWeek(weekIndex));
+    return getRange(this.getWeekCount()).map(weekIndex => this.renderWeek(weekIndex));
   }
 
   renderMonthLabels() {
-    if (!this.props.showMonthLabels) {
+    if (!this.latestProps.showMonthLabels) {
       return null;
     }
-    const weekRange = range(this.getWeekCount() - 1);  // don't render for last week, because label will be cut off
+    const weekRange = getRange(this.getWeekCount() - 1); // don't render for last week, because label will be cut off
     return weekRange.map((weekIndex) => {
       const endOfWeek = shiftDate(this.getStartDateWithEmptyDays(), (weekIndex + 1) * DAYS_IN_WEEK);
       const [x, y] = this.getMonthLabelCoordinates(weekIndex);
       return (endOfWeek.getDate() >= 1 && endOfWeek.getDate() <= DAYS_IN_WEEK) ? (
-        <text
-          key={weekIndex}
-          x={x}
-          y={y}
-        >
-          {MONTH_LABELS[endOfWeek.getMonth()]}
+        <text key={weekIndex} x={x} y={y} className={`${CSS_PSEDUO_NAMESPACE}month-label`}>
+          {this.latestProps.monthLabels[endOfWeek.getMonth()]}
+        </text>
+      ) : null;
+    });
+  }
+
+  renderWeekdayLabels() {
+    if (!this.latestProps.showWeekdayLabels) {
+      return null;
+    }
+    return this.latestProps.weekdayLabels.map((weekdayLabel, dayIndex) => {
+      const [x, y] = this.getWeekdayLabelCoordinates(dayIndex);
+      const cssClasses = `${this.latestProps.horizontal ? '' : `${CSS_PSEDUO_NAMESPACE}small-text`} ${CSS_PSEDUO_NAMESPACE}weekday-label`;
+      // eslint-disable-next-line no-bitwise
+      return dayIndex & 1 ? (
+        <text key={`${x}${y}`} x={x} y={y} className={cssClasses}>
+          {weekdayLabel}
         </text>
       ) : null;
     });
@@ -239,15 +311,15 @@ class CalendarHeatmap extends React.Component {
 
   render() {
     return (
-      <svg
-        className="react-calendar-heatmap"
-        viewBox={this.getViewBox()}
-      >
-        <g transform={this.getTransformForMonthLabels()}>
+      <svg className="react-calendar-heatmap" viewBox={this.getViewBox()}>
+        <g transform={this.getTransformForMonthLabels()} className={`${CSS_PSEDUO_NAMESPACE}month-labels`}>
           {this.renderMonthLabels()}
         </g>
-        <g transform={this.getTransformForAllWeeks()}>
+        <g transform={this.getTransformForAllWeeks()} className={`${CSS_PSEDUO_NAMESPACE}all-weeks`}>
           {this.renderAllWeeks()}
+        </g>
+        <g transform={this.getTransformForWeekdayLabels()} className={`${CSS_PSEDUO_NAMESPACE}weekday-labels`}>
+          {this.renderWeekdayLabels()}
         </g>
       </svg>
     );
@@ -255,31 +327,38 @@ class CalendarHeatmap extends React.Component {
 }
 
 CalendarHeatmap.propTypes = {
-  values: PropTypes.arrayOf(             // array of objects with date and arbitrary metadata
-    PropTypes.shape({
-      date: PropTypes.oneOfType([PropTypes.string, PropTypes.number, PropTypes.instanceOf(Date)]).isRequired,
-    }).isRequired
-  ).isRequired,
-  numDays: PropTypes.number,             // number of days back from endDate to show
-  endDate: PropTypes.oneOfType([PropTypes.string, PropTypes.number, PropTypes.instanceOf(Date)]),  // end of date range
-  gutterSize: PropTypes.number,          // size of space between squares
-  horizontal: PropTypes.bool,            // whether to orient horizontally or vertically
-  showMonthLabels: PropTypes.bool,       // whether to show month labels
-  showOutOfRangeDays: PropTypes.bool,    // whether to render squares for extra days in week after endDate, and before start date
-  tooltipDataAttrs: PropTypes.oneOfType([PropTypes.object, PropTypes.func]),    // data attributes to add to square for setting 3rd party tooltips, e.g. { 'data-toggle': 'tooltip' } for bootstrap tooltips
-  titleForValue: PropTypes.func,         // function which returns title text for value
-  classForValue: PropTypes.func,         // function which returns html class for value
-  onClick: PropTypes.func,               // callback function when a square is clicked
-  transformDayElement: PropTypes.func,    // function to further transform the svg element for a single day
+  values: PropTypes.arrayOf(PropTypes.shape({
+    date: PropTypes.oneOfType([PropTypes.string, PropTypes.number, PropTypes.instanceOf(Date)]).isRequired,
+  }).isRequired).isRequired, // array of objects with date and arbitrary metadata
+  numDays: PropTypes.number, // number of days back from endDate to show
+  startDate: PropTypes.oneOfType([PropTypes.string, PropTypes.number, PropTypes.instanceOf(Date)]), // start of date range
+  endDate: PropTypes.oneOfType([PropTypes.string, PropTypes.number, PropTypes.instanceOf(Date)]), // end of date range
+  gutterSize: PropTypes.number, // size of space between squares
+  horizontal: PropTypes.bool, // whether to orient horizontally or vertically
+  showMonthLabels: PropTypes.bool, // whether to show month labels
+  showWeekdayLabels: PropTypes.bool, // whether to show weekday labels
+  showOutOfRangeDays: PropTypes.bool, // whether to render squares for extra days in week after endDate, and before start date
+  tooltipDataAttrs: PropTypes.oneOfType([PropTypes.object, PropTypes.func]), // data attributes to add to square for setting 3rd party tooltips, e.g. { 'data-toggle': 'tooltip' } for bootstrap tooltips
+  titleForValue: PropTypes.func, // function which returns title text for value
+  classForValue: PropTypes.func, // function which returns html class for value
+  monthLabels: PropTypes.arrayOf(PropTypes.string), // An array with 12 strings representing the text from janurary to december
+  weekdayLabels: PropTypes.arrayOf(PropTypes.string), // An array with 7 strings representing the text from Sun to Sat
+  onClick: PropTypes.func, // callback function when a square is clicked
+  onMouseOver: PropTypes.func, // callback function when mouse pointer is over a square
+  onMouseLeave: PropTypes.func, // callback function when mouse pointer is left a square
+  transformDayElement: PropTypes.func, // function to further transform the svg element for a single day
 };
 
 CalendarHeatmap.defaultProps = {
-  numDays: 200,
+  startDate: dateNDaysAgo(200),
   endDate: new Date(),
   gutterSize: 1,
   horizontal: true,
   showMonthLabels: true,
+  showWeekdayLabels: false,
   showOutOfRangeDays: false,
+  monthLabels: MONTH_LABELS,
+  weekdayLabels: DAY_LABELS,
   classForValue: value => (value ? 'color-filled' : 'color-empty'),
 };