+++ CHANGELOG.md
@@ -1,34 +1,72 @@
-## 1.2.1
+* Fix tooltips in demo page (thanks @kaacun!)
+
+## v1.6.3
+
+* Fix occurences where startDate & endDate were not properly converted (thanks @kenkoooo!)
+
+## v1.6.2
+
+* Use latest props when updating value cache (thanks @gitname!)
+* Add cross-env so builds work on Windows (thanks @gitname!)
+* Expand test coverage (thanks @gitname!)
+
+## v1.6.1
+
+* Make prop-types a dependency, not a devDependency
+
+## v1.6.0
+
+* Add css class name hooks to most inner html elements (thanks @andreysaleba!)
+
+## v1.5.0
+
+* Add startDate prop & deprecate numDays prop (thanks @andreysaleba!)
+
+## v1.4.1
+
+* Allow React 16 as a peerDependency
+
+## v1.4.0
+
+* Add weekday label display (thanks @andreysaleba!)
+* Add title element to tooltip (thanks @andreysaleba!)
+* Add onMouseLeave, onMouseOver props (thanks @andreysaleba!)
+
+## v1.3.0
+
+* Allow customizeable month strings (thanks @otakustay!)
+
+## v1.2.1
 
 * Update React to `v15.6.1` & add `prop-types` package (thanks @easingthemes!)
 
-## 1.2.0
+## v1.2.0
 
 * Add transformDayElement prop to improve flexibility (thanks @otakustay!)
 
-## 1.1.0
+## v1.1.0
 
 * Allow tooltipDataAttrs to be a function rather than an object (thanks @RSO!)
 
-## 1.0.0
+## v1.0.0
 
 * Improve tooltip support with tooltipDataAttrs prop, allowing usage of e.g. bootstrap tooltips
 * titleForValue now sets a square's title attribute instead of setting svg `<title>` element
 
-## 0.4.3
+## v0.4.3
 
 * Build configuration updates
 
-## 0.4.2
+## v0.4.2
 
 * Allow endDate to be string/milliseconds
 * Don't show titles by default
 
-## 0.4.1
+## v0.4.1
 
 * Fix broken package issue
 * Stop loading CSS through webpack
 
-## 0.4.0
+## v0.4.0
 
 * Add `horizontal` prop, which allows switching between horizontal and vertical orientations